% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mthd_author.R
\name{author-SpecificProject}
\alias{author-SpecificProject}
\alias{author,SpecificProject-method}
\title{Author and save files to local or remote locations}
\usage{
\S4method{author}{SpecificProject}(
  x,
  fl_nm_1L_chr = "I_ALL_OUTPUT_",
  path_1L_chr = NA_character_,
  type_1L_chr = "results",
  what_1L_chr = "public"
)
}
\arguments{
\item{x}{An object of class SpecificProject}

\item{fl_nm_1L_chr}{File name (a character vector of length one), Default: 'I_ALL_OUTPUT_'}

\item{path_1L_chr}{Path (a character vector of length one), Default: 'NA'}

\item{type_1L_chr}{Type (a character vector of length one), Default: 'results'}

\item{what_1L_chr}{What (a character vector of length one), Default: 'public'}
}
\description{
author method applied to SpecificProject
}
