% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mthd_authorData.R
\name{authorData-SpecificMixed}
\alias{authorData-SpecificMixed}
\alias{authorData,SpecificMixed-method}
\title{Author and document datasets}
\usage{
\S4method{authorData}{SpecificMixed}(x, title_1L_chr = "An R model object", what_1L_chr = "Shareable")
}
\arguments{
\item{x}{An object of class SpecificMixed}

\item{title_1L_chr}{Title (a character vector of length one), Default: 'An R model object'}

\item{what_1L_chr}{What (a character vector of length one), Default: 'Shareable'}
}
\value{
x (An object of class SpecificMixed)
}
\description{
authorData method applied to SpecificMixed
}
