% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mthd_depict.R
\name{depict-SpecificSynopsis}
\alias{depict-SpecificSynopsis}
\alias{depict,SpecificSynopsis-method}
\alias{depict-SpecificProject}
\alias{depict,SpecificProject-method}
\title{Depict (plot) features of a ready4 framework module (or sub-module)}
\usage{
\S4method{depict}{SpecificSynopsis}(
  x,
  base_height = 13,
  base_size_1L_dbl = 30,
  depnt_var_desc_1L_chr = NA_character_,
  labels_chr = c("A", "B", "C", "D"),
  label_x_1L_dbl = 0.2,
  label_y_1L_dbl = 0.9,
  label_size_1L_dbl = 30,
  mdl_idxs_int = 1:2,
  timepoint_old_nms_chr = NA_character_,
  timepoint_new_nms_chr = NA_character_,
  use_png_fls_1L_lgl = F,
  what_1L_chr = "composite_mdl",
  write_1L_lgl = F
)

\S4method{depict}{SpecificProject}(
  x,
  mdl_idxs_int = NULL,
  output_type_1L_chr = "HTML",
  plt_indcs_int = NULL
)
}
\arguments{
\item{x}{An object of class SpecificProject}

\item{base_height}{PARAM_DESCRIPTION, Default: 13}

\item{base_size_1L_dbl}{Base size (a double vector of length one), Default: 30}

\item{depnt_var_desc_1L_chr}{Dependent variable description (a character vector of length one), Default: 'NA'}

\item{labels_chr}{Labels (a character vector), Default: c("A", "B", "C", "D")}

\item{label_x_1L_dbl}{Label x (a double vector of length one), Default: 0.2}

\item{label_y_1L_dbl}{Label y (a double vector of length one), Default: 0.9}

\item{label_size_1L_dbl}{Label size (a double vector of length one), Default: 30}

\item{mdl_idxs_int}{Model indices (an integer vector), Default: NULL}

\item{timepoint_old_nms_chr}{Timepoint old names (a character vector), Default: 'NA'}

\item{timepoint_new_nms_chr}{Timepoint new names (a character vector), Default: 'NA'}

\item{use_png_fls_1L_lgl}{Use png files (a logical vector of length one), Default: F}

\item{what_1L_chr}{What (a character vector of length one), Default: 'composite_mdl'}

\item{write_1L_lgl}{Write (a logical vector of length one), Default: F}

\item{output_type_1L_chr}{Output type (a character vector of length one), Default: 'HTML'}

\item{plt_indcs_int}{Plot indcs (an integer vector), Default: NULL}
}
\value{
Plot (a plot)
}
\description{
depict method applied to SpecificSynopsis

depict method applied to SpecificProject
}
