% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mthd_exhibit.R
\name{exhibit-SpecificParameters}
\alias{exhibit-SpecificParameters}
\alias{exhibit,SpecificParameters-method}
\title{Exhibit features of a class instance by printing to console}
\usage{
\S4method{exhibit}{SpecificParameters}(
  x,
  captions_chr = character(0),
  method_chr = c("pearson", "spearman"),
  mkdn_tbl_refs_chr = NULL,
  output_type_1L_chr = "HTML",
  type_1L_chr = "correlation",
  timepoints_int = NA_integer_
)
}
\arguments{
\item{x}{An object of class SpecificParameters}

\item{captions_chr}{Captions (a character vector), Default: character(0)}

\item{method_chr}{Method (a character vector), Default: c("pearson", "spearman")}

\item{mkdn_tbl_refs_chr}{Markdown table references (a character vector), Default: NULL}

\item{output_type_1L_chr}{Output type (a character vector of length one), Default: 'HTML'}

\item{type_1L_chr}{Type (a character vector of length one), Default: 'correlation'}

\item{timepoints_int}{Timepoints (an integer vector), Default: NA}
}
\description{
exhibit method applied to SpecificParameters
}
