% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fn_get.R
\name{get_brms_mdl}
\alias{get_brms_mdl}
\title{Get bayesian regression models model}
\usage{
get_brms_mdl(outp_smry_ls, mdl_nm_1L_chr)
}
\arguments{
\item{outp_smry_ls}{Output summary (a list)}

\item{mdl_nm_1L_chr}{Model name (a character vector of length one)}
}
\value{
Bayesian regression models (a model)
}
\description{
get_brms_mdl() is a Get function that retrieves a pre-existing data object from memory, local file system or online repository. Specifically, this function implements an algorithm to get bayesian regression models model. Function argument outp_smry_ls specifies the where to look for the required object. The function returns Bayesian regression models (a model).
}
\keyword{internal}
