% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fn_get.R
\name{get_cndt_mdls}
\alias{get_cndt_mdls}
\title{Get candidate models}
\usage{
get_cndt_mdls(
  filter_1L_lgl = T,
  mdl_short_nms_chr = NA_character_,
  mdl_types_lup = NULL
)
}
\arguments{
\item{filter_1L_lgl}{Filter (a logical vector of length one), Default: T}

\item{mdl_short_nms_chr}{Model short names (a character vector), Default: 'NA'}

\item{mdl_types_lup}{Model types (a lookup table), Default: NULL}
}
\value{
Candidate models (a lookup table)
}
\description{
get_cndt_mdls() is a Get function that retrieves a pre-existing data object from memory, local file system or online repository. Specifically, this function implements an algorithm to get candidate models. Function argument filter_1L_lgl specifies the where to look for the required object. The function returns Candidate models (a lookup table).
}
\keyword{internal}
