% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fn_get.R
\name{get_signft_covars}
\alias{get_signft_covars}
\title{Get significant covariates}
\usage{
get_signft_covars(
  mdls_with_covars_smry_tb,
  covar_var_nms_chr,
  what_1L_chr = "any",
  X_Ready4useDyad = NULL
)
}
\arguments{
\item{mdls_with_covars_smry_tb}{Models with covariates summary (a tibble)}

\item{covar_var_nms_chr}{Covariate variable names (a character vector)}

\item{what_1L_chr}{What (a character vector of length one), Default: 'any'}

\item{X_Ready4useDyad}{PARAM_DESCRIPTION, Default: NULL}
}
\value{
Signt covariates (a character vector)
}
\description{
get_signft_covars() is a Get function that retrieves a pre-existing data object from memory, local file system or online repository. Specifically, this function implements an algorithm to get significant covariates. Function argument mdls_with_covars_smry_tb specifies the where to look for the required object. The function returns Signt covariates (a character vector).
}
