% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fn_make.R
\name{make_csnl_example_predrs}
\alias{make_csnl_example_predrs}
\title{Make cross-sectional example predictors}
\usage{
make_csnl_example_predrs()
}
\value{
Predictors (a ready4 S3)
}
\description{
make_csnl_example_predrs() is a Make function that creates a new R object. Specifically, this function implements an algorithm to make cross-sectional example predictors. The function returns Predictors (a ready4 S3).
}
\keyword{internal}
