% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fn_make.R
\name{make_incld_mdl_paths}
\alias{make_incld_mdl_paths}
\title{Make included model paths}
\usage{
make_incld_mdl_paths(outp_smry_ls)
}
\arguments{
\item{outp_smry_ls}{Output summary (a list)}
}
\value{
Included model paths (a character vector)
}
\description{
make_incld_mdl_paths() is a Make function that creates a new R object. Specifically, this function implements an algorithm to make included model paths. The function returns Included model paths (a character vector).
}
\keyword{internal}
