% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fn_make.R
\name{make_ms_smry_tbls_ls}
\alias{make_ms_smry_tbls_ls}
\title{Make manuscript summary tables list}
\usage{
make_ms_smry_tbls_ls(
  outp_smry_ls,
  mdls_smry_tbls_ls,
  covars_mdls_ls,
  descv_tbls_ls,
  nbr_of_digits_1L_int = 2L
)
}
\arguments{
\item{outp_smry_ls}{Output summary (a list)}

\item{mdls_smry_tbls_ls}{Models summary tables (a list)}

\item{covars_mdls_ls}{Covariates models (a list)}

\item{descv_tbls_ls}{Descriptive tables (a list)}

\item{nbr_of_digits_1L_int}{Number of digits (an integer vector of length one), Default: 2}
}
\value{
Manuscript summary tables (a list)
}
\description{
make_ms_smry_tbls_ls() is a Make function that creates a new R object. Specifically, this function implements an algorithm to make manuscript summary tables list. The function returns Manuscript summary tables (a list).
}
\keyword{internal}
