% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/C3_specific_models.R
\name{make_new_specific_models}
\alias{make_new_specific_models}
\title{make new specific models Candidate models lookup table}
\usage{
make_new_specific_models(x)
}
\arguments{
\item{x}{A prototype for the Candidate models lookup table}
}
\value{
An unvalidated instance of the Candidate models lookup table
}
\description{
Create a new unvalidated instance of the Candidate models lookup table
}
\details{
Candidate models lookup table
}
\keyword{internal}
