% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/C3_specific_predictors.R
\name{make_new_specific_predictors}
\alias{make_new_specific_predictors}
\title{make new specific predictors Candidate predictors lookup table}
\usage{
make_new_specific_predictors(x)
}
\arguments{
\item{x}{A prototype for the Candidate predictors lookup table}
}
\value{
An unvalidated instance of the Candidate predictors lookup table
}
\description{
Create a new unvalidated instance of the Candidate predictors lookup table
}
\details{
Candidate predictors lookup table
}
\keyword{internal}
