% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fn_make.R
\name{make_plot_fn_and_args_ls}
\alias{make_plot_fn_and_args_ls}
\title{Make plot function and arguments list}
\usage{
make_plot_fn_and_args_ls(
  type_1L_chr,
  depnt_var_desc_1L_chr,
  args_ls = NULL,
  base_size_1L_dbl = 11,
  brms_mdl = NULL,
  correspondences_lup = NULL,
  depnt_var_nm_1L_chr = NULL,
  new_var_nm_1L_chr = NA_character_,
  predn_type_1L_chr = NULL,
  round_var_nm_1L_chr = NULL,
  sd_dbl = NA_real_,
  seed_1L_dbl = 23456,
  sfx_1L_chr = " from table",
  table_predn_mdl = NULL,
  tfmn_1L_chr = "NTF",
  tfd_data_tb = NULL,
  utl_min_val_1L_dbl = -1,
  x_lbl_1L_chr = NA_character_,
  y_lbl_1L_chr = NA_character_
)
}
\arguments{
\item{type_1L_chr}{Type (a character vector of length one)}

\item{depnt_var_desc_1L_chr}{Dependent variable description (a character vector of length one)}

\item{args_ls}{Arguments (a list), Default: NULL}

\item{base_size_1L_dbl}{Base size (a double vector of length one), Default: 11}

\item{brms_mdl}{Bayesian regression models (a model), Default: NULL}

\item{correspondences_lup}{Correspondences (a lookup table), Default: NULL}

\item{depnt_var_nm_1L_chr}{Dependent variable name (a character vector of length one), Default: NULL}

\item{new_var_nm_1L_chr}{New variable name (a character vector of length one), Default: 'NA'}

\item{predn_type_1L_chr}{Prediction type (a character vector of length one), Default: NULL}

\item{round_var_nm_1L_chr}{Round variable name (a character vector of length one), Default: NULL}

\item{sd_dbl}{Standard deviation (a double vector), Default: NA}

\item{seed_1L_dbl}{Seed (a double vector of length one), Default: 23456}

\item{sfx_1L_chr}{Suffix (a character vector of length one), Default: ' from table'}

\item{table_predn_mdl}{Table prediction (a model), Default: NULL}

\item{tfmn_1L_chr}{Transformation (a character vector of length one), Default: 'NTF'}

\item{tfd_data_tb}{Transformed data (a tibble), Default: NULL}

\item{utl_min_val_1L_dbl}{Utility minimum value (a double vector of length one), Default: -1}

\item{x_lbl_1L_chr}{X label (a character vector of length one), Default: 'NA'}

\item{y_lbl_1L_chr}{Y label (a character vector of length one), Default: 'NA'}
}
\value{
Plot function and arguments (a list)
}
\description{
make_plot_fn_and_args_ls() is a Make function that creates a new R object. Specifically, this function implements an algorithm to make plot function and arguments list. The function returns Plot function and arguments (a list).
}
\keyword{internal}
