% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fn_make.R
\name{make_results_ls_spine}
\alias{make_results_ls_spine}
\title{Make results list spine}
\usage{
make_results_ls_spine(
  output_format_ls = NULL,
  params_ls_ls = NULL,
  path_params_ls = NULL,
  study_descs_ls,
  fn_ls = NULL,
  include_idx_int = NULL,
  outp_smry_ls = NULL,
  var_nm_change_lup = NULL
)
}
\arguments{
\item{output_format_ls}{Output format (a list), Default: NULL}

\item{params_ls_ls}{Parameters (a list of lists), Default: NULL}

\item{path_params_ls}{Path parameters (a list), Default: NULL}

\item{study_descs_ls}{Study descriptions (a list)}

\item{fn_ls}{Function list (a list of functions), Default: NULL}

\item{include_idx_int}{Include index (an integer vector), Default: NULL}

\item{outp_smry_ls}{Output summary (a list), Default: NULL}

\item{var_nm_change_lup}{Variable name change (a lookup table), Default: NULL}
}
\value{
Spine of results (a list)
}
\description{
make_results_ls_spine() is a Make function that creates a new R object. Specifically, this function implements an algorithm to make results list spine. The function returns Spine of results (a list).
}
\keyword{internal}
