% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fn_make.R
\name{make_scaling_fctr_dbl}
\alias{make_scaling_fctr_dbl}
\title{Make scaling factor double vector}
\usage{
make_scaling_fctr_dbl(outp_smry_ls)
}
\arguments{
\item{outp_smry_ls}{Output summary (a list)}
}
\value{
Scaling factor (a double vector)
}
\description{
make_scaling_fctr_dbl() is a Make function that creates a new R object. Specifically, this function implements an algorithm to make scaling factor double vector. The function returns Scaling factor (a double vector).
}
\keyword{internal}
