% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fn_make.R
\name{make_shareable_mdl}
\alias{make_shareable_mdl}
\title{Make shareable model}
\usage{
make_shareable_mdl(
  fake_ds_tb,
  mdl_smry_tb,
  x_ready4use_dictionary,
  control_1L_chr = NA_character_,
  depnt_var_nm_1L_chr = "utl_total_w",
  id_var_nm_1L_chr = "fkClientID",
  mdl_type_1L_chr = "OLS_CLL",
  mdl_types_lup = NULL,
  seed_1L_int = 12345L,
  start_1L_chr = NA_character_,
  tfmn_1L_chr = "CLL"
)
}
\arguments{
\item{fake_ds_tb}{Fake dataset (a tibble)}

\item{mdl_smry_tb}{Model summary (a tibble)}

\item{x_ready4use_dictionary}{PARAM_DESCRIPTION}

\item{control_1L_chr}{Control (a character vector of length one), Default: 'NA'}

\item{depnt_var_nm_1L_chr}{Dependent variable name (a character vector of length one), Default: 'utl_total_w'}

\item{id_var_nm_1L_chr}{Identity variable name (a character vector of length one), Default: 'fkClientID'}

\item{mdl_type_1L_chr}{Model type (a character vector of length one), Default: 'OLS_CLL'}

\item{mdl_types_lup}{Model types (a lookup table), Default: NULL}

\item{seed_1L_int}{Seed (an integer vector of length one), Default: 12345}

\item{start_1L_chr}{Start (a character vector of length one), Default: 'NA'}

\item{tfmn_1L_chr}{Transformation (a character vector of length one), Default: 'CLL'}
}
\value{
Model (a model)
}
\description{
make_shareable_mdl() is a Make function that creates a new R object. Specifically, this function implements an algorithm to make shareable model. The function returns Model (a model).
}
