% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mthd_manufacture.R
\name{manufacture-SpecificMixed}
\alias{manufacture-SpecificMixed}
\alias{manufacture,SpecificMixed-method}
\alias{manufacture-SpecificProject}
\alias{manufacture,SpecificProject-method}
\title{Manufacture a (non ready4 framework) object}
\usage{
\S4method{manufacture}{SpecificMixed}(
  x,
  a_Ready4useRepos = NULL,
  scndry_anlys_params_ls = NULL,
  what_1L_chr = "input_params_ls"
)

\S4method{manufacture}{SpecificProject}(x, what_1L_chr = "ds_descvs_ls")
}
\arguments{
\item{x}{An object of class SpecificProject}

\item{a_Ready4useRepos}{PARAM_DESCRIPTION, Default: NULL}

\item{scndry_anlys_params_ls}{Secondary analysis parameters (a list), Default: NULL}

\item{what_1L_chr}{What (a character vector of length one), Default: 'ds_descvs_ls'}
}
\value{
Object (an output object of multiple potential types)

Object (an output object of multiple potential types)
}
\description{
manufacture method applied to SpecificMixed

manufacture method applied to SpecificProject
}
