% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mthd_renew.R
\name{renew-SpecificProject}
\alias{renew-SpecificProject}
\alias{renew,SpecificProject-method}
\title{Renew (update) an object}
\usage{
\S4method{renew}{SpecificProject}(x, new_val_xx, type_1L_chr = "results", what_1L_chr = "prefd_mdls")
}
\arguments{
\item{x}{An object of class SpecificProject}

\item{new_val_xx}{New value (an output object of multiple potential types)}

\item{type_1L_chr}{Type (a character vector of length one), Default: 'results'}

\item{what_1L_chr}{What (a character vector of length one), Default: 'prefd_mdls'}
}
\value{
x (An object of class SpecificProject)
}
\description{
renew method applied to SpecificProject
}
