% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/pkg_specific.R, R/specific-package.R
\docType{package}
\name{specific-package}
\alias{specific}
\alias{specific-package}
\title{specific: Specify Models to Solve Inverse Problems}
\description{
\if{html}{\figure{logo.png}{options: style='float: right' alt='logo' width='120'}}

Tools to standardise the formalising and testing of models to solve inverse problems. Designed for use in models developed with the ready4 framework (https://www.ready4-dev.com/). This development version of the specific package has been made available as part of the process of testing and documenting the package. If you have any questions, please contact the authors (matthew.hamilton1@monash.edu).

\if{html}{\figure{logo.png}{options: style='float: right' alt='logo' width='120'}}

Tools to standardise the formalising and testing of models to solve inverse problems. Designed for use in models developed with the ready4 framework (https://www.ready4-dev.com/). This development version of the specific package has been made available as part of the process of testing and documenting the package. If you have any questions, please contact the authors (matthew.hamilton1@monash.edu).
}
\seealso{
Useful links:
\itemize{
  \item \url{https://ready4-dev.github.io/specific/}
  \item \url{https://github.com/ready4-dev/specific}
  \item \url{https://www.ready4-dev.com/}
}


Useful links:
\itemize{
  \item \url{https://ready4-dev.github.io/specific/}
  \item \url{https://github.com/ready4-dev/specific}
  \item \url{https://www.ready4-dev.com/}
}

}
\author{
\strong{Maintainer}: Matthew Hamilton \email{matthew.hamilton1@monash.edu} (\href{https://orcid.org/0000-0001-7407-9194}{ORCID})

Authors:
\itemize{
  \item Caroline Gao \email{caroline.gao@orygen.org.au} (\href{https://orcid.org/0000-0002-0987-2759}{ORCID})
}

Other contributors:
\itemize{
  \item Orygen [copyright holder, funder]
  \item Headspace [funder]
  \item National Health and Medical Research Council [funder]
}

}
\keyword{internal}
