% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/C3_specific_models.R
\name{specific_models}
\alias{specific_models}
\alias{make_pt_specific_models}
\alias{is_specific_models}
\title{Candidate models lookup table}
\usage{
specific_models(x = make_pt_specific_models())

make_pt_specific_models(
  short_name_chr = character(0),
  long_name_chr = character(0),
  control_chr = character(0),
  family_chr = character(0),
  fn_chr = character(0),
  start_chr = character(0),
  predn_type_chr = character(0),
  tfmn_chr = character(0),
  tfmn_for_bnml_lgl = logical(0),
  fixed_acronym_chr = character(0),
  mixed_acronym_chr = character(0),
  mixed_type_chr = character(0),
  with_chr = character(0)
)

is_specific_models(x)
}
\arguments{
\item{x}{An object of any type}

\item{short_name_chr}{Short name (a character vector), Default: character(0)}

\item{long_name_chr}{Long name (a character vector), Default: character(0)}

\item{control_chr}{Control (a character vector), Default: character(0)}

\item{family_chr}{Family (a character vector), Default: character(0)}

\item{fn_chr}{Function (a character vector), Default: character(0)}

\item{start_chr}{Start (a character vector), Default: character(0)}

\item{predn_type_chr}{Prediction type (a character vector), Default: character(0)}

\item{tfmn_chr}{Transformation (a character vector), Default: character(0)}

\item{tfmn_for_bnml_lgl}{Transformation for binomial (a logical vector), Default: logical(0)}

\item{fixed_acronym_chr}{Fixed acronym (a character vector), Default: character(0)}

\item{mixed_acronym_chr}{Mixed acronym (a character vector), Default: character(0)}

\item{mixed_type_chr}{Mixed type (a character vector), Default: character(0)}

\item{with_chr}{With (a character vector), Default: character(0)}
}
\value{
A validated instance of the Candidate models lookup table

A prototype for Candidate models lookup table

A logical value, TRUE if a valid instance of the Candidate models lookup table
}
\description{
Create a new valid instance of the Candidate models lookup table
}
\details{
Candidate models lookup table
}
