% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/C3_specific_predictors.R
\name{specific_predictors}
\alias{specific_predictors}
\alias{make_pt_specific_predictors}
\alias{is_specific_predictors}
\title{Candidate predictors lookup table}
\usage{
specific_predictors(x = make_pt_specific_predictors())

make_pt_specific_predictors(
  short_name_chr = character(0),
  long_name_chr = character(0),
  min_val_dbl = numeric(0),
  max_val_dbl = numeric(0),
  class_chr = character(0),
  increment_dbl = numeric(0),
  class_fn_chr = character(0),
  mdl_scaling_dbl = numeric(0),
  covariate_lgl = logical(0)
)

is_specific_predictors(x)
}
\arguments{
\item{x}{An object of any type}

\item{short_name_chr}{Short name (a character vector), Default: character(0)}

\item{long_name_chr}{Long name (a character vector), Default: character(0)}

\item{min_val_dbl}{Minimum value (a double vector), Default: numeric(0)}

\item{max_val_dbl}{Maximum value (a double vector), Default: numeric(0)}

\item{class_chr}{Class (a character vector), Default: character(0)}

\item{increment_dbl}{Increment (a double vector), Default: numeric(0)}

\item{class_fn_chr}{Class function (a character vector), Default: character(0)}

\item{mdl_scaling_dbl}{Model scaling (a double vector), Default: numeric(0)}

\item{covariate_lgl}{Covariate (a logical vector), Default: logical(0)}
}
\value{
A validated instance of the Candidate predictors lookup table

A prototype for Candidate predictors lookup table

A logical value, TRUE if a valid instance of the Candidate predictors lookup table
}
\description{
Create a new valid instance of the Candidate predictors lookup table
}
\details{
Candidate predictors lookup table
}
