% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/C4_SpecificConverter.R
\docType{class}
\name{SpecificConverter-class}
\alias{SpecificConverter-class}
\alias{SpecificConverter}
\title{SpecificConverter}
\description{
Container for seed objects used for creating SpecificModels modules.
}
\section{Slots}{

\describe{
\item{\code{a_ScorzProfile}}{(an instance of the ScorzProfile class)}

\item{\code{dissemination_1L_chr}}{Dissemination (a character vector of length one)}
}}

