% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/C4_SpecificModels.R
\docType{class}
\name{SpecificModels-class}
\alias{SpecificModels-class}
\alias{SpecificModels}
\title{SpecificModels}
\description{
Modelling project dataset, input parameters and model comparison results.
}
\section{Slots}{

\describe{
\item{\code{a_YouthvarsProfile}}{(an instance of the YouthvarsProfile class)}

\item{\code{b_SpecificParameters}}{(an instance of the SpecificParameters class)}

\item{\code{c_SpecificResults}}{(an instance of the SpecificResults class)}

\item{\code{paths_chr}}{Paths (a character vector)}

\item{\code{dissemination_1L_chr}}{Dissemination (a character vector of length one)}
}}

