% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/C4_SpecificPredictors.R
\docType{class}
\name{SpecificPredictors-class}
\alias{SpecificPredictors-class}
\alias{SpecificPredictors}
\title{SpecificPredictors}
\description{
Modelling project dataset, input parameters and predictor comparison results.
}
\section{Slots}{

\describe{
\item{\code{a_YouthvarsProfile}}{(an instance of the YouthvarsProfile class)}

\item{\code{b_SpecificParameters}}{(an instance of the SpecificParameters class)}

\item{\code{c_SpecificResults}}{(an instance of the SpecificResults class)}

\item{\code{dissemination_1L_chr}}{Dissemination (a character vector of length one)}
}}

