% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/C4_SpecificPrivate.R
\docType{class}
\name{SpecificPrivate-class}
\alias{SpecificPrivate-class}
\alias{SpecificPrivate}
\title{SpecificPrivate}
\description{
Analysis outputs not intended for public dissemination.
}
\section{Slots}{

\describe{
\item{\code{private_outp_ls}}{Private output (a list)}

\item{\code{dissemination_1L_chr}}{Dissemination (a character vector of length one)}
}}

