% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/C4_SpecificProject.R
\docType{class}
\name{SpecificProject-class}
\alias{SpecificProject-class}
\alias{SpecificProject}
\title{SpecificProject}
\description{
Modelling project input parameters and results output.
}
\section{Slots}{

\describe{
\item{\code{a_YouthvarsProfile}}{(an instance of the YouthvarsProfile class)}

\item{\code{b_SpecificParameters}}{(an instance of the SpecificParameters class)}

\item{\code{c_SpecificResults}}{(an instance of the SpecificResults class)}

\item{\code{dissemination_1L_chr}}{Dissemination (a character vector of length one)}
}}

