% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/C4_SpecificResults.R
\docType{class}
\name{SpecificResults-class}
\alias{SpecificResults-class}
\alias{SpecificResults}
\title{SpecificResults}
\description{
Analysis results.
}
\section{Slots}{

\describe{
\item{\code{a_SpecificShareable}}{(an instance of the SpecificShareable class)}

\item{\code{b_SpecificPrivate}}{(an instance of the SpecificPrivate class)}

\item{\code{dissemination_1L_chr}}{Dissemination (a character vector of length one)}
}}

