% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/C4_SpecificShareable.R
\docType{class}
\name{SpecificShareable-class}
\alias{SpecificShareable-class}
\alias{SpecificShareable}
\title{SpecificShareable}
\description{
Analysis outputs intended for public dissemination.
}
\section{Slots}{

\describe{
\item{\code{shareable_outp_ls}}{Shareable output (a list)}

\item{\code{dissemination_1L_chr}}{Dissemination (a character vector of length one)}
}}

