% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fn_transform.R
\name{transform_tbls_for_covar_nms}
\alias{transform_tbls_for_covar_nms}
\title{Transform tables for covariate names}
\usage{
transform_tbls_for_covar_nms(results_ls)
}
\arguments{
\item{results_ls}{Results (a list)}
}
\value{
Results (a list)
}
\description{
transform_tbls_for_covar_nms() is a Transform function that edits an object in such a way that core object attributes - e.g. shape, dimensions, elements, type - are altered. Specifically, this function implements an algorithm to transform tables for covariate names. Function argument results_ls specifies the object to be updated. The function returns Results (a list).
}
\keyword{internal}
