% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fn_transform.R
\name{transform_tbls_for_csnl_mdls}
\alias{transform_tbls_for_csnl_mdls}
\title{Transform tables for cross-sectional models}
\usage{
transform_tbls_for_csnl_mdls(results_ls)
}
\arguments{
\item{results_ls}{Results (a list)}
}
\value{
Results (a list)
}
\description{
transform_tbls_for_csnl_mdls() is a Transform function that edits an object in such a way that core object attributes - e.g. shape, dimensions, elements, type - are altered. Specifically, this function implements an algorithm to transform tables for cross-sectional models. Function argument results_ls specifies the object to be updated. The function returns Results (a list).
}
\keyword{internal}
