% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/C3_specific_models.R
\name{validate_specific_models}
\alias{validate_specific_models}
\title{validate specific models Candidate models lookup table}
\usage{
validate_specific_models(x)
}
\arguments{
\item{x}{An unvalidated instance of the Candidate models lookup table}
}
\value{
A prototpe for Candidate models lookup table
}
\description{
Validate an instance of the Candidate models lookup table
}
\details{
Candidate models lookup table
}
\keyword{internal}
