% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/C3_specific_predictors.R
\name{validate_specific_predictors}
\alias{validate_specific_predictors}
\title{validate specific predictors Candidate predictors lookup table}
\usage{
validate_specific_predictors(x)
}
\arguments{
\item{x}{An unvalidated instance of the Candidate predictors lookup table}
}
\value{
A prototpe for Candidate predictors lookup table
}
\description{
Validate an instance of the Candidate predictors lookup table
}
\details{
Candidate predictors lookup table
}
\keyword{internal}
