% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fn_write.R
\name{write_mdl_smry_rprt}
\alias{write_mdl_smry_rprt}
\title{Write model summary report}
\usage{
write_mdl_smry_rprt(
  input_params_ls = NULL,
  abstract_args_ls = NULL,
  dv_ds_nm_and_url_chr = NULL,
  rprt_lup = NULL,
  rcrd_nm_1L_chr = "AAA_RPRT_WRTNG_MTH",
  rprt_nm_1L_chr = "AAA_TTU_MDL_CTG",
  start_at_int = c(2, 1),
  use_shareable_mdls_1L_lgl = F
)
}
\arguments{
\item{input_params_ls}{Input parameters (a list), Default: NULL}

\item{abstract_args_ls}{Abstract arguments (a list), Default: NULL}

\item{dv_ds_nm_and_url_chr}{Dataverse dataset name and url (a character vector), Default: NULL}

\item{rprt_lup}{Report (a lookup table), Default: NULL}

\item{rcrd_nm_1L_chr}{Record name (a character vector of length one), Default: 'AAA_RPRT_WRTNG_MTH'}

\item{rprt_nm_1L_chr}{Report name (a character vector of length one), Default: 'AAA_TTU_MDL_CTG'}

\item{start_at_int}{Start at (an integer vector), Default: c(2, 1)}

\item{use_shareable_mdls_1L_lgl}{Use shareable models (a logical vector of length one), Default: F}
}
\value{
Input parameters (a list)
}
\description{
write_mdl_smry_rprt() is a Write function that writes a file to a specified local directory. Specifically, this function implements an algorithm to write model summary report. The function returns Input parameters (a list).
}
\keyword{internal}
