% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fn_write.R
\name{write_predr_and_covars_cmprsn}
\alias{write_predr_and_covars_cmprsn}
\title{Write predictor and covariates comparison}
\usage{
write_predr_and_covars_cmprsn(
  scored_data_tb,
  bl_tb,
  consent_1L_chr = "",
  consent_indcs_int = 1L,
  depnt_var_min_val_1L_dbl = numeric(0),
  ds_smry_ls,
  mdl_smry_ls,
  options_chr = c("Y", "N"),
  output_data_dir_1L_chr,
  seed_1L_int = 1234,
  signft_covars_cdn_1L_chr = "any"
)
}
\arguments{
\item{scored_data_tb}{Scored data (a tibble)}

\item{bl_tb}{Baseline (a tibble)}

\item{consent_1L_chr}{Consent (a character vector of length one), Default: ''}

\item{consent_indcs_int}{Consent indices (an integer vector), Default: 1}

\item{depnt_var_min_val_1L_dbl}{Dependent variable minimum value (a double vector of length one), Default: numeric(0)}

\item{ds_smry_ls}{Dataset summary (a list)}

\item{mdl_smry_ls}{Model summary (a list)}

\item{options_chr}{Options (a character vector), Default: c("Y", "N")}

\item{output_data_dir_1L_chr}{Output data directory (a character vector of length one)}

\item{seed_1L_int}{Seed (an integer vector of length one), Default: 1234}

\item{signft_covars_cdn_1L_chr}{Significant covariates condition (a character vector of length one), Default: 'any'}
}
\value{
Predictor and covariates comparison (a list)
}
\description{
write_predr_and_covars_cmprsn() is a Write function that writes a file to a specified local directory. Specifically, this function implements an algorithm to write predictor and covariates comparison. The function returns Predictor and covariates comparison (a list).
}
