% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fn_write.R
\name{write_predr_and_mdl_tstng_results}
\alias{write_predr_and_mdl_tstng_results}
\title{Write predictor and model testing results}
\usage{
write_predr_and_mdl_tstng_results(
  scored_data_tb,
  combinations_1L_lgl = F,
  consent_1L_chr = "",
  consent_indcs_int = 1L,
  depnt_var_max_val_1L_dbl = 0.99999,
  depnt_var_min_val_1L_dbl = 1e-05,
  ds_smry_ls,
  existing_predrs_ls = NULL,
  max_nbr_of_covars_1L_int = integer(0),
  mdl_smry_ls,
  session_data_ls,
  options_chr = c("Y", "N"),
  output_data_dir_1L_chr,
  seed_1L_int = 1234,
  signft_covars_cdn_1L_chr = "any"
)
}
\arguments{
\item{scored_data_tb}{Scored data (a tibble)}

\item{combinations_1L_lgl}{Combinations (a logical vector of length one), Default: F}

\item{consent_1L_chr}{Consent (a character vector of length one), Default: ''}

\item{consent_indcs_int}{Consent indices (an integer vector), Default: 1}

\item{depnt_var_max_val_1L_dbl}{Dependent variable maximum value (a double vector of length one), Default: 0.99999}

\item{depnt_var_min_val_1L_dbl}{Dependent variable minimum value (a double vector of length one), Default: 1e-05}

\item{ds_smry_ls}{Dataset summary (a list)}

\item{existing_predrs_ls}{Existing predictors (a list), Default: NULL}

\item{max_nbr_of_covars_1L_int}{Maximum number of covariates (an integer vector of length one), Default: integer(0)}

\item{mdl_smry_ls}{Model summary (a list)}

\item{session_data_ls}{Session data (a list)}

\item{options_chr}{Options (a character vector), Default: c("Y", "N")}

\item{output_data_dir_1L_chr}{Output data directory (a character vector of length one)}

\item{seed_1L_int}{Seed (an integer vector of length one), Default: 1234}

\item{signft_covars_cdn_1L_chr}{Significant covariates condition (a character vector of length one), Default: 'any'}
}
\value{
Output summary (a list)
}
\description{
write_predr_and_mdl_tstng_results() is a Write function that writes a file to a specified local directory. Specifically, this function implements an algorithm to write predictor and model testing results. The function returns Output summary (a list).
}
\keyword{internal}
