% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fn_write.R
\name{write_shareable_mdls}
\alias{write_shareable_mdls}
\title{Write shareable models}
\usage{
write_shareable_mdls(
  outp_smry_ls,
  consent_1L_chr = "",
  consent_indcs_int = 1L,
  depnt_var_min_val_1L_dbl = numeric(0),
  new_dir_nm_1L_chr = "G_Shareable",
  options_chr = c("Y", "N"),
  shareable_title_detail_1L_chr = "",
  write_mdls_to_dv_1L_lgl = F
)
}
\arguments{
\item{outp_smry_ls}{Output summary (a list)}

\item{consent_1L_chr}{Consent (a character vector of length one), Default: ''}

\item{consent_indcs_int}{Consent indices (an integer vector), Default: 1}

\item{depnt_var_min_val_1L_dbl}{Dependent variable minimum value (a double vector of length one), Default: numeric(0)}

\item{new_dir_nm_1L_chr}{New directory name (a character vector of length one), Default: 'G_Shareable'}

\item{options_chr}{Options (a character vector), Default: c("Y", "N")}

\item{shareable_title_detail_1L_chr}{Shareable title detail (a character vector of length one), Default: ''}

\item{write_mdls_to_dv_1L_lgl}{Write models to dataverse (a logical vector of length one), Default: F}
}
\value{
Output summary (a list)
}
\description{
write_shareable_mdls() is a Write function that writes a file to a specified local directory. Specifically, this function implements an algorithm to write shareable models. The function returns Output summary (a list).
}
