% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fn_write.R
\name{write_ts_mdl_plts}
\alias{write_ts_mdl_plts}
\title{Write time series model plots}
\usage{
write_ts_mdl_plts(
  brms_mdl,
  mdl_nm_1L_chr,
  path_to_write_to_1L_chr,
  tfd_data_tb,
  args_ls = NULL,
  consent_1L_chr = "",
  consent_indcs_int = 1L,
  depnt_var_desc_1L_chr = "Utility score",
  depnt_var_nm_1L_chr = "utl_total_w",
  height_dbl = c(rep(6, 2), rep(5, 8)),
  options_chr = c("Y", "N"),
  predn_type_1L_chr = NULL,
  round_var_nm_1L_chr = "round",
  rsl_dbl = rep(300, 10),
  sd_dbl = NA_real_,
  seed_1L_dbl = 23456,
  sfx_1L_chr = " from table",
  table_predn_mdl = NULL,
  tfmn_1L_chr = "NTF",
  units_1L_chr = "in",
  utl_min_val_1L_dbl = -1,
  width_dbl = c(rep(6, 2), rep(6, 8))
)
}
\arguments{
\item{brms_mdl}{Bayesian regression models (a model)}

\item{mdl_nm_1L_chr}{Model name (a character vector of length one)}

\item{path_to_write_to_1L_chr}{Path to write to (a character vector of length one)}

\item{tfd_data_tb}{Transformed data (a tibble)}

\item{args_ls}{Arguments (a list), Default: NULL}

\item{consent_1L_chr}{Consent (a character vector of length one), Default: ''}

\item{consent_indcs_int}{Consent indices (an integer vector), Default: 1}

\item{depnt_var_desc_1L_chr}{Dependent variable description (a character vector of length one), Default: 'Utility score'}

\item{depnt_var_nm_1L_chr}{Dependent variable name (a character vector of length one), Default: 'utl_total_w'}

\item{height_dbl}{Height (a double vector), Default: c(rep(6, 2), rep(5, 8))}

\item{options_chr}{Options (a character vector), Default: c("Y", "N")}

\item{predn_type_1L_chr}{Prediction type (a character vector of length one), Default: NULL}

\item{round_var_nm_1L_chr}{Round variable name (a character vector of length one), Default: 'round'}

\item{rsl_dbl}{Resolution (a double vector), Default: rep(300, 10)}

\item{sd_dbl}{Standard deviation (a double vector), Default: NA}

\item{seed_1L_dbl}{Seed (a double vector of length one), Default: 23456}

\item{sfx_1L_chr}{Suffix (a character vector of length one), Default: ' from table'}

\item{table_predn_mdl}{Table prediction (a model), Default: NULL}

\item{tfmn_1L_chr}{Transformation (a character vector of length one), Default: 'NTF'}

\item{units_1L_chr}{Units (a character vector of length one), Default: 'in'}

\item{utl_min_val_1L_dbl}{Utility minimum value (a double vector of length one), Default: -1}

\item{width_dbl}{Width (a double vector), Default: c(rep(6, 2), rep(6, 8))}
}
\value{
Model plots paths (a list)
}
\description{
write_ts_mdl_plts() is a Write function that writes a file to a specified local directory. Specifically, this function implements an algorithm to write time series model plots. The function returns Model plots paths (a list).
}
\keyword{internal}
