% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mthd_depict.R
\name{depict-SpecificProject}
\alias{depict-SpecificProject}
\alias{depict,SpecificProject-method}
\title{Depict features of an instance of a class by generating a plot}
\usage{
\S4method{depict}{SpecificProject}(
  x,
  mdl_idcs_int = NULL,
  output_type_1L_chr = "HTML",
  plt_indcs_int = NULL
)
}
\arguments{
\item{x}{An object of class SpecificProject}

\item{mdl_idcs_int}{Model idcs (an integer vector), Default: NULL}

\item{output_type_1L_chr}{Output type (a character vector of length one), Default: 'HTML'}

\item{plt_indcs_int}{Plot indcs (an integer vector), Default: NULL}
}
\description{
depict method applied to SpecificProject
}
