% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mthd_exhibit.R
\name{exhibit.specific_models}
\alias{exhibit.specific_models}
\alias{exhibit,specific_models-method}
\alias{exhibit.specific_predictors}
\alias{exhibit,specific_predictors-method}
\alias{exhibit-SpecificProject}
\alias{exhibit,SpecificProject-method}
\title{Exhibit (print to console) an object}
\usage{
exhibit.specific_models(
  x,
  caption_1L_chr = NULL,
  mkdn_tbl_ref_1L_chr = NULL,
  output_type_1L_chr = "HTML",
  use_lbls_as_col_nms_1L_lgl = T
)

\S4method{exhibit}{specific_models}(
  x,
  caption_1L_chr = NULL,
  mkdn_tbl_ref_1L_chr = NULL,
  output_type_1L_chr = "HTML",
  use_lbls_as_col_nms_1L_lgl = T
)

exhibit.specific_predictors(
  x,
  caption_1L_chr = NULL,
  mkdn_tbl_ref_1L_chr = NULL,
  output_type_1L_chr = "HTML",
  use_lbls_as_col_nms_1L_lgl = T
)

\S4method{exhibit}{specific_predictors}(
  x,
  caption_1L_chr = NULL,
  mkdn_tbl_ref_1L_chr = NULL,
  output_type_1L_chr = "HTML",
  use_lbls_as_col_nms_1L_lgl = T
)

\S4method{exhibit}{SpecificProject}(
  x,
  captions_chr = character(0),
  method_chr = c("pearson", "spearman"),
  mkdn_tbl_refs_chr = NULL,
  output_type_1L_chr = "HTML",
  timepoints_int = NA_integer_,
  type_1L_chr = "data",
  what_1L_chr = "correlation",
  ...
)
}
\arguments{
\item{x}{An object of class SpecificProject}

\item{caption_1L_chr}{Caption (a character vector of length one), Default: NULL}

\item{mkdn_tbl_ref_1L_chr}{Markdown table reference (a character vector of length one), Default: NULL}

\item{output_type_1L_chr}{Output type (a character vector of length one), Default: 'HTML'}

\item{use_lbls_as_col_nms_1L_lgl}{Use labels as column names (a logical vector of length one), Default: T}

\item{captions_chr}{Captions (a character vector), Default: character(0)}

\item{method_chr}{Method (a character vector), Default: c("pearson", "spearman")}

\item{mkdn_tbl_refs_chr}{Markdown table references (a character vector), Default: NULL}

\item{timepoints_int}{Timepoints (an integer vector), Default: NA}

\item{type_1L_chr}{Type (a character vector of length one), Default: 'data'}

\item{what_1L_chr}{What (a character vector of length one), Default: 'correlation'}

\item{...}{Additional arguments}
}
\description{
exhibit.specific_models() is an exhibit method that exhibits features of a class instance by printing to console. This method is implemented for the Candidate models lookup table The function is called for its side effects and does not return a value.

exhibit.specific_predictors() is an exhibit method that exhibits features of a class instance by printing to console. This method is implemented for the Candidate predictors lookup table The function is called for its side effects and does not return a value.

exhibit method applied to SpecificProject
}
