% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mthd_investigate.R
\name{investigate-SpecificModels}
\alias{investigate-SpecificModels}
\alias{investigate,SpecificModels-method}
\alias{investigate-SpecificMixed}
\alias{investigate,SpecificMixed-method}
\alias{investigate-SpecificPredictors}
\alias{investigate,SpecificPredictors-method}
\alias{investigate-SpecificFixed}
\alias{investigate,SpecificFixed-method}
\title{Investigate solutions to an inverse problem}
\usage{
\S4method{investigate}{SpecificModels}(x, depnt_var_max_val_1L_dbl = Inf, session_ls = NULL)

\S4method{investigate}{SpecificMixed}(
  x,
  backend_1L_chr = "cmdstanr",
  new_dir_nm_1L_chr = "F_TS_Mdls",
  scndry_anlys_params_ls = NULL
)

\S4method{investigate}{SpecificPredictors}(x)

\S4method{investigate}{SpecificFixed}(x)
}
\arguments{
\item{x}{An object of class SpecificFixed}

\item{depnt_var_max_val_1L_dbl}{Dependent variable maximum value (a double vector of length one), Default: Inf}

\item{session_ls}{Session (a list), Default: NULL}

\item{backend_1L_chr}{Backend (a character vector of length one), Default: 'cmdstanr'}

\item{new_dir_nm_1L_chr}{New directory name (a character vector of length one), Default: 'F_TS_Mdls'}

\item{scndry_anlys_params_ls}{Secondary analysis parameters (a list), Default: NULL}
}
\value{
X (Modelling project dataset, input parameters and predictor comparison results.)

x (An object of class SpecificMixed)

X (Modelling project dataset, input parameters and complete fixed models results.)

X (Modelling project dataset, input parameters and complete mixed models results.)
}
\description{
investigate method applied to SpecificModels

investigate method applied to SpecificMixed

investigate method applied to SpecificPredictors

investigate method applied to SpecificFixed
}
