% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mthd_investigate.R
\name{investigate-SpecificModels}
\alias{investigate-SpecificModels}
\alias{investigate,SpecificModels-method}
\alias{investigate-SpecificMixed}
\alias{investigate,SpecificMixed-method}
\alias{investigate-SpecificPredictors}
\alias{investigate,SpecificPredictors-method}
\alias{investigate-SpecificFixed}
\alias{investigate,SpecificFixed-method}
\title{Investigate data stored in an instance of a class by executing an algorithm to solve an inverse problem}
\usage{
\S4method{investigate}{SpecificModels}(x)

\S4method{investigate}{SpecificMixed}(x, backend_1L_chr = "cmdstanr", new_dir_nm_1L_chr = "F_TS_Mdls")

\S4method{investigate}{SpecificPredictors}(x)

\S4method{investigate}{SpecificFixed}(x)
}
\arguments{
\item{x}{An object of class SpecificFixed}

\item{backend_1L_chr}{Backend (a character vector of length one), Default: 'cmdstanr'}

\item{new_dir_nm_1L_chr}{New directory name (a character vector of length one), Default: 'F_TS_Mdls'}
}
\value{
x (An object of class SpecificMixed)

X (Modelling project dataset, input parameters and complete fixed models results.)

X (Modelling project dataset, input parameters and complete mixed models results.)
}
\description{
investigate method applied to SpecificModels

investigate method applied to SpecificMixed

investigate method applied to SpecificPredictors

investigate method applied to SpecificFixed
}
