% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fn_make.R
\name{make_inner_loop_mdl_smry}
\alias{make_inner_loop_mdl_smry}
\title{Make inner loop model summary}
\usage{
make_inner_loop_mdl_smry(
  idx_1L_int,
  data_tb,
  mdl_nms_ls,
  mdl_smry_dir_1L_chr,
  mdl_types_lup,
  predictors_lup,
  predr_vars_nms_ls,
  depnt_var_min_val_1L_dbl = numeric(0),
  depnt_var_nm_1L_chr = "utl_total_w",
  id_var_nm_1L_chr = "fkClientID",
  round_var_nm_1L_chr = "round",
  round_bl_val_1L_chr = "Baseline",
  utl_min_val_1L_dbl = -1,
  backend_1L_chr = getOption("brms.backend", "rstan"),
  iters_1L_int = 4000L,
  seed_1L_int = 1000L,
  prior_ls = NULL,
  control_ls = NULL
)
}
\arguments{
\item{idx_1L_int}{Index (an integer vector of length one)}

\item{data_tb}{Data (a tibble)}

\item{mdl_nms_ls}{Model names (a list)}

\item{mdl_smry_dir_1L_chr}{Model summary directory (a character vector of length one)}

\item{mdl_types_lup}{Model types (a lookup table)}

\item{predictors_lup}{Predictors (a lookup table)}

\item{predr_vars_nms_ls}{Predictor variables names (a list)}

\item{depnt_var_min_val_1L_dbl}{Dependent variable minimum value (a double vector of length one), Default: numeric(0)}

\item{depnt_var_nm_1L_chr}{Dependent variable name (a character vector of length one), Default: 'utl_total_w'}

\item{id_var_nm_1L_chr}{Identity variable name (a character vector of length one), Default: 'fkClientID'}

\item{round_var_nm_1L_chr}{Round variable name (a character vector of length one), Default: 'round'}

\item{round_bl_val_1L_chr}{Round baseline value (a character vector of length one), Default: 'Baseline'}

\item{utl_min_val_1L_dbl}{Utility minimum value (a double vector of length one), Default: -1}

\item{backend_1L_chr}{Backend (a character vector of length one), Default: getOption("brms.backend", "rstan")}

\item{iters_1L_int}{Iterations (an integer vector of length one), Default: 4000}

\item{seed_1L_int}{Seed (an integer vector of length one), Default: 1000}

\item{prior_ls}{Prior (a list), Default: NULL}

\item{control_ls}{Control (a list), Default: NULL}
}
\value{
Models summary (a tibble)
}
\description{
make_inner_loop_mdl_smry() is a Make function that creates a new R object. Specifically, this function implements an algorithm to make inner loop model summary. The function returns Models summary (a tibble).
}
\keyword{internal}
