% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mthd_manufacture.R
\name{manufacture-SpecificModels}
\alias{manufacture-SpecificModels}
\alias{manufacture,SpecificModels-method}
\title{Manufacture a (non ready4 framework) object}
\usage{
\S4method{manufacture}{SpecificModels}(x, what_1L_chr = "ds_descvs_ls")
}
\arguments{
\item{x}{An object of class SpecificModels}

\item{what_1L_chr}{What (a character vector of length one), Default: 'ds_descvs_ls'}
}
\value{
Object (an output object of multiple potential types)
}
\description{
manufacture method applied to SpecificModels
}
