% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mthd_manufacture.R
\name{manufacture-SpecificResults}
\alias{manufacture-SpecificResults}
\alias{manufacture,SpecificResults-method}
\alias{manufacture-SpecificProject}
\alias{manufacture,SpecificProject-method}
\alias{manufacture-SpecificSynopsis}
\alias{manufacture,SpecificSynopsis-method}
\title{Manufacture a new object}
\usage{
\S4method{manufacture}{SpecificResults}(x, what_1L_chr = "indexed_shareable", ...)

\S4method{manufacture}{SpecificProject}(
  x,
  what_1L_chr = "ds_descvs_ls",
  scndry_anlys_params_ls = NULL,
  ...
)

\S4method{manufacture}{SpecificSynopsis}(
  x,
  consent_1L_chr = "",
  depnt_var_min_val_1L_dbl = numeric(0),
  depnt_var_nms_chr = NA_character_,
  make_cmpst_plt_1L_lgl = F,
  scndry_anlys_params_ls = NULL,
  version_1L_chr = "",
  what_1L_chr = "input_params_ls",
  ...
)
}
\arguments{
\item{x}{An object of class SpecificSynopsis}

\item{what_1L_chr}{What (a character vector of length one), Default: 'input_params_ls'}

\item{...}{Additional arguments}

\item{scndry_anlys_params_ls}{Secondary analysis parameters (a list), Default: NULL}

\item{consent_1L_chr}{Consent (a character vector of length one), Default: ''}

\item{depnt_var_min_val_1L_dbl}{Dependent variable minimum value (a double vector of length one), Default: numeric(0)}

\item{depnt_var_nms_chr}{Dependent variable names (a character vector), Default: 'NA'}

\item{make_cmpst_plt_1L_lgl}{Make composite plot (a logical vector of length one), Default: F}

\item{version_1L_chr}{Version (a character vector of length one), Default: ''}
}
\value{
Object (an output object of multiple potential types)

Object (an output object of multiple potential types)

Object (an output object of multiple potential types)
}
\description{
manufacture method applied to SpecificResults

manufacture method applied to SpecificProject

manufacture method applied to SpecificSynopsis
}
