% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mthd_metamorphose.R
\name{metamorphose-SpecificMixed}
\alias{metamorphose-SpecificMixed}
\alias{metamorphose,SpecificMixed-method}
\alias{metamorphose-SpecificConverter}
\alias{metamorphose,SpecificConverter-method}
\title{Metamorphose from one ready4 framework module (or sub-module) class to another}
\usage{
\S4method{metamorphose}{SpecificMixed}(x, to_1L_chr = "SpecificSynopsis", rmd_fl_nms_ls = NULL)

\S4method{metamorphose}{SpecificConverter}(x, paths_chr = NA_character_)
}
\arguments{
\item{x}{An object of class SpecificConverter}

\item{to_1L_chr}{To (a character vector of length one), Default: 'SpecificSynopsis'}

\item{rmd_fl_nms_ls}{R Markdown file names (a list), Default: NULL}

\item{paths_chr}{Paths (a character vector), Default: 'NA'}
}
\value{
Y (a ready4 S4)

X (Modelling project dataset, input parameters and model comparison results.)
}
\description{
metamorphose method applied to SpecificMixed

metamorphose method applied to SpecificConverter
}
