% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mthd_metamorphose.R
\name{metamorphose-SpecificConverter}
\alias{metamorphose-SpecificConverter}
\alias{metamorphose,SpecificConverter-method}
\title{Metamorphose from one ready4 framework module (or sub-module) class to another}
\usage{
\S4method{metamorphose}{SpecificConverter}(x, paths_chr = NA_character_)
}
\arguments{
\item{x}{An object of class SpecificConverter}

\item{paths_chr}{Paths (a character vector), Default: 'NA'}
}
\value{
X (Modelling project dataset, input parameters and model comparison results.)
}
\description{
metamorphose method applied to SpecificConverter
}
