% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mthd_ratify.R
\name{ratify-SpecificModels}
\alias{ratify-SpecificModels}
\alias{ratify,SpecificModels-method}
\title{Ratify that a dataset meets validity criteria}
\usage{
\S4method{ratify}{SpecificModels}(
  x,
  class_fn_1L_chr = "as.numeric",
  prototype_lup = NULL,
  scndry_anlys_params_ls = NULL,
  ...
)
}
\arguments{
\item{x}{An object of class SpecificModels}

\item{class_fn_1L_chr}{Class function (a character vector of length one), Default: 'as.numeric'}

\item{prototype_lup}{Prototype (a lookup table), Default: NULL}

\item{scndry_anlys_params_ls}{Secondary analysis parameters (a list), Default: NULL}

\item{...}{Additional arguments}
}
\value{
x (An object of class SpecificModels)
}
\description{
ratify method applied to SpecificModels
}
