% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mthd_renew.R
\name{renew.specific_models}
\alias{renew.specific_models}
\alias{renew,specific_models-method}
\alias{renew.specific_predictors}
\alias{renew,specific_predictors-method}
\alias{renew-SpecificProject}
\alias{renew,SpecificProject-method}
\alias{renew-SpecificMixed}
\alias{renew,SpecificMixed-method}
\title{Renew values in a dataset}
\usage{
renew.specific_models(
  x,
  short_name_chr = NA_character_,
  long_name_chr = NA_character_,
  control_chr = NA_character_,
  family_chr = NA_character_,
  fn_chr = NA_character_,
  start_chr = NA_character_,
  predn_type_chr = NA_character_,
  tfmn_chr = NA_character_,
  tfmn_for_bnml_lgl = NA,
  fixed_acronym_chr = NA_character_,
  mixed_acronym_chr = NA_character_,
  mixed_type_chr = NA_character_,
  with_chr = NA_character_,
  filter_cdn_1L_chr = NA_character_,
  new_cases_r3 = NULL,
  new_ready4_dict_r3 = deprecated(),
  slice_indcs_int = NA_integer_
)

\S4method{renew}{specific_models}(
  x,
  short_name_chr = NA_character_,
  long_name_chr = NA_character_,
  control_chr = NA_character_,
  family_chr = NA_character_,
  fn_chr = NA_character_,
  start_chr = NA_character_,
  predn_type_chr = NA_character_,
  tfmn_chr = NA_character_,
  tfmn_for_bnml_lgl = NA,
  fixed_acronym_chr = NA_character_,
  mixed_acronym_chr = NA_character_,
  mixed_type_chr = NA_character_,
  with_chr = NA_character_,
  filter_cdn_1L_chr = NA_character_,
  new_cases_r3 = NULL,
  new_ready4_dict_r3 = deprecated(),
  slice_indcs_int = NA_integer_
)

renew.specific_predictors(
  x,
  short_name_chr = NA_character_,
  long_name_chr = NA_character_,
  min_val_dbl = NA_real_,
  max_val_dbl = NA_real_,
  class_chr = NA_character_,
  increment_dbl = NA_real_,
  class_fn_chr = NA_character_,
  mdl_scaling_dbl = NA_real_,
  covariate_lgl = NA,
  filter_cdn_1L_chr = NA_character_,
  new_cases_r3 = NULL,
  new_ready4_dict_r3 = deprecated(),
  slice_indcs_int = NA_integer_
)

\S4method{renew}{specific_predictors}(
  x,
  short_name_chr = NA_character_,
  long_name_chr = NA_character_,
  min_val_dbl = NA_real_,
  max_val_dbl = NA_real_,
  class_chr = NA_character_,
  increment_dbl = NA_real_,
  class_fn_chr = NA_character_,
  mdl_scaling_dbl = NA_real_,
  covariate_lgl = NA,
  filter_cdn_1L_chr = NA_character_,
  new_cases_r3 = NULL,
  new_ready4_dict_r3 = deprecated(),
  slice_indcs_int = NA_integer_
)

\S4method{renew}{SpecificProject}(x, new_val_xx, type_1L_chr = "results", what_1L_chr = "prefd_mdls", ...)

\S4method{renew}{SpecificMixed}(
  x,
  new_val_xx = NULL,
  a_Ready4useRepos,
  type_1L_chr = "results",
  what_1L_chr = "dv_ls",
  ...
)
}
\arguments{
\item{x}{An object of class SpecificMixed}

\item{short_name_chr}{Short name (a character vector), Default: 'NA'}

\item{long_name_chr}{Long name (a character vector), Default: 'NA'}

\item{control_chr}{Control (a character vector), Default: 'NA'}

\item{family_chr}{Family (a character vector), Default: 'NA'}

\item{fn_chr}{Function (a character vector), Default: 'NA'}

\item{start_chr}{Start (a character vector), Default: 'NA'}

\item{predn_type_chr}{Prediction type (a character vector), Default: 'NA'}

\item{tfmn_chr}{Transformation (a character vector), Default: 'NA'}

\item{tfmn_for_bnml_lgl}{Transformation for binomial (a logical vector), Default: NA}

\item{fixed_acronym_chr}{Fixed acronym (a character vector), Default: 'NA'}

\item{mixed_acronym_chr}{Mixed acronym (a character vector), Default: 'NA'}

\item{mixed_type_chr}{Mixed type (a character vector), Default: 'NA'}

\item{with_chr}{With (a character vector), Default: 'NA'}

\item{filter_cdn_1L_chr}{Filter condition (a character vector of length one), Default: 'NA'}

\item{new_cases_r3}{New cases (a ready4 S3), Default: NULL}

\item{new_ready4_dict_r3}{New ready4 dictionary (a ready4 S3), Default: deprecated()}

\item{slice_indcs_int}{Slice indices (an integer vector), Default: NA}

\item{min_val_dbl}{Minimum value (a double vector), Default: NA}

\item{max_val_dbl}{Maximum value (a double vector), Default: NA}

\item{class_chr}{Class (a character vector), Default: 'NA'}

\item{increment_dbl}{Increment (a double vector), Default: NA}

\item{class_fn_chr}{Class function (a character vector), Default: 'NA'}

\item{mdl_scaling_dbl}{Model scaling (a double vector), Default: NA}

\item{covariate_lgl}{Covariate (a logical vector), Default: NA}

\item{new_val_xx}{New value (an output object of multiple potential types), Default: NULL}

\item{type_1L_chr}{Type (a character vector of length one), Default: 'results'}

\item{what_1L_chr}{What (a character vector of length one), Default: 'dv_ls'}

\item{...}{Additional arguments}

\item{a_Ready4useRepos}{PARAM_DESCRIPTION}
}
\value{
x (An object)

x (An object)

x (An object of class SpecificProject)

x (An object of class SpecificMixed)
}
\description{
renew.specific_models() is a renew method that renews an instance of a class by updating it with new data. This method is implemented for the Candidate models lookup table The function is called for its side effects and does not return a value.

renew.specific_predictors() is a renew method that renews an instance of a class by updating it with new data. This method is implemented for the Candidate predictors lookup table The function is called for its side effects and does not return a value.

renew method applied to SpecificProject

renew method applied to SpecificMixed
}
