% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mthd_share.R
\name{share-SpecificProject}
\alias{share-SpecificProject}
\alias{share,SpecificProject-method}
\title{Share data contained in an instance of a class via an online repository}
\usage{
\S4method{share}{SpecificProject}(x, fl_nm_1L_chr, repos_Ready4useRepos)
}
\arguments{
\item{x}{An object of class SpecificProject}

\item{fl_nm_1L_chr}{File name (a character vector of length one)}

\item{repos_Ready4useRepos}{PARAM_DESCRIPTION}
}
\description{
share method applied to SpecificProject
}
