% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/C4_SpecificParameters.R
\docType{class}
\name{SpecificParameters-class}
\alias{SpecificParameters-class}
\alias{SpecificParameters}
\title{SpecificParameters}
\description{
Input parameters that specify candidate models to be explored.
}
\section{Slots}{

\describe{
\item{\code{candidate_mdls_chr}}{Candidate models (a character vector)}

\item{\code{candidate_mdls_lup}}{Candidate models (a lookup table)}

\item{\code{candidate_mdl_pfcs_chr}}{Candidate model pfcs (a character vector)}

\item{\code{candidate_predrs_chr}}{Candidate predictors (a character vector)}

\item{\code{candidate_covars_chr}}{Candidate covariates (a character vector)}

\item{\code{control_ls}}{Control (a list)}

\item{\code{depnt_var_nm_1L_chr}}{Dependent variable name (a character vector of length one)}

\item{\code{depnt_var_min_max_dbl}}{Dependent variable minimum maximum (a double vector)}

\item{\code{folds_1L_int}}{Folds (an integer vector of length one)}

\item{\code{iters_1L_int}}{Iterations (an integer vector of length one)}

\item{\code{max_mdl_runs_1L_int}}{Maximum model runs (an integer vector of length one)}

\item{\code{prior_ls}}{Prior (a list)}

\item{\code{seed_1L_int}}{Seed (an integer vector of length one)}

\item{\code{predictors_lup}}{Predictors (a lookup table)}

\item{\code{dissemination_1L_chr}}{Dissemination (a character vector of length one)}
}}

