% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fn_write.R
\name{write_scndry_analysis}
\alias{write_scndry_analysis}
\title{Write secondary analysis}
\usage{
write_scndry_analysis(
  valid_params_ls_ls,
  candidate_covar_nms_chr,
  path_params_ls,
  reference_1L_int,
  backend_1L_chr = "cmdstanr",
  candidate_predrs_chr = NULL,
  new_dir_nm_1L_chr = "F_TS_Mdls",
  predictors_lup = NULL,
  prefd_covars_chr = NA_character_
)
}
\arguments{
\item{valid_params_ls_ls}{Valid parameters (a list of lists)}

\item{candidate_covar_nms_chr}{Candidate covariate names (a character vector)}

\item{path_params_ls}{Path parameters (a list)}

\item{reference_1L_int}{Reference (an integer vector of length one)}

\item{backend_1L_chr}{Backend (a character vector of length one), Default: 'cmdstanr'}

\item{candidate_predrs_chr}{Candidate predictors (a character vector), Default: NULL}

\item{new_dir_nm_1L_chr}{New directory name (a character vector of length one), Default: 'F_TS_Mdls'}

\item{predictors_lup}{Predictors (a lookup table), Default: NULL}

\item{prefd_covars_chr}{Preferred covariates (a character vector), Default: 'NA'}
}
\value{
Output summary (a list)
}
\description{
write_scndry_analysis() is a Write function that writes a file to a specified local directory. Specifically, this function implements an algorithm to write secondary analysis. The function returns Output summary (a list).
}
\keyword{internal}
