% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fn_write.R
\name{write_secondary_analyses}
\alias{write_secondary_analyses}
\title{Write secondary analyses}
\usage{
write_secondary_analyses(
  input_params_ls,
  backend_1L_chr = "cmdstanr",
  new_dir_nm_1L_chr = "F_TS_Mdls"
)
}
\arguments{
\item{input_params_ls}{Input parameters (a list)}

\item{backend_1L_chr}{Backend (a character vector of length one), Default: 'cmdstanr'}

\item{new_dir_nm_1L_chr}{New directory name (a character vector of length one), Default: 'F_TS_Mdls'}
}
\value{
Results (a list)
}
\description{
write_secondary_analyses() is a Write function that writes a file to a specified local directory. Specifically, this function implements an algorithm to write secondary analyses. The function returns Results (a list).
}
\keyword{internal}
